# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/bots:
    get:
      tags: [Bots]
      summary: List created bots and their configurations
      operationId: listBots
      x-astrbot-scope: bot
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Bots]
      summary: Create a bot from configuration
      operationId: createBot
      x-astrbot-scope: bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/bots/by-id:
    get:
      tags: [Bots]
      summary: Get a bot configuration by query ID
      operationId: getBotById
      x-astrbot-scope: bot
      parameters:
        - name: bot_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Bots]
      summary: Replace a bot configuration by body ID
      operationId: updateBotById
      x-astrbot-scope: bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bot_id, config]
              properties:
                bot_id:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Bots]
      summary: Delete a bot by query ID
      operationId: deleteBotById
      x-astrbot-scope: bot
      parameters:
        - name: bot_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/bots/enabled:
    patch:
      tags: [Bots]
      summary: Enable or disable a bot by body ID
      operationId: setBotEnabledById
      x-astrbot-scope: bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bot_id, enabled]
              properties:
                bot_id:
                  type: string
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/bots/test:
    post:
      tags: [Bots]
      summary: Test a bot registration or connection by body ID
      operationId: testBotById
      x-astrbot-scope: bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bot_id]
              properties:
                bot_id:
                  type: string
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/bots/{bot_id}:
    get:
      tags: [Bots]
      summary: Get a bot configuration
      operationId: getBot
      x-astrbot-scope: bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Bots]
      summary: Replace a bot configuration
      operationId: updateBot
      x-astrbot-scope: bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Bots]
      summary: Delete a bot
      operationId: deleteBot
      x-astrbot-scope: bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/bots/{bot_id}/enabled:
    patch:
      tags: [Bots]
      summary: Enable or disable a bot
      operationId: setBotEnabled
      x-astrbot-scope: bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/bots/{bot_id}/test:
    post:
      tags: [Bots]
      summary: Test a bot registration or connection
      operationId: testBot
      x-astrbot-scope: bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    BotId:
      name: bot_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    BotConfigRequest:
      type: object
      required: [type, config]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          description: Platform adapter type, such as aiocqhttp, telegram, lark.
        enabled:
          type: boolean
          default: true
        config:
          description: "Wangshangliao supports reply_private (boolean, default true) and reply_groups (group ID to boolean map, omitted entries default true). Group replies still require mentions."
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
...
    EnabledPatch:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
      additionalProperties: false
...